// app/src/generated-api/models/Vehicle.ts
export interface Vehicle {
    /**
     * ID único gerado pelo sistema.
     * @type {string}
     * @memberof Vehicle
     */
    id: string;
    /**
     * Timestamp de criação do registro.
     * @type {Date}
     * @memberof Vehicle
     */
    createdAt?: Date;
    /**
     * Placa do veículo. Formato Mercosul (LLLNLNN).
     * @type {string}
     * @memberof Vehicle
     */
    licensePlate: string;
    /**
     * Modelo e nome do veículo.
     * @type {string}
     * @memberof Vehicle
     */
    model: string;
    /**
     * Fabricante do veículo.
     * @type {string}
     * @memberof Vehicle
     */
    vehicleManufacturer?: string;
    /**
     * Ano do Modelo do veículo.
     * @type {number}
     * @memberof Vehicle
     */
    modelYear?: number;
    /**
     * Ano de fabricação do veículo.
     * @type {number}
     * @memberof Vehicle
     */
    manufacturerYear?: number;
    /**
     * Número do RENAVAN do veículo.
     * @type {string}
     * @memberof Vehicle
     */
    renavan?: string;
    /**
     * Em caso de veículo de transporte, o número de passageiros.
     * @type {number}
     * @memberof Vehicle
     */
    passengerNumber?: number;
    /**
     * Número do motor do veículo.
     * @type {string}
     * @memberof Vehicle
     */
    motorNumber?: string;
    /**
     * Número do chassis do veículo.
     * @type {string}
     * @memberof Vehicle
     */
    chassisNumber?: string;
    /**
     * Número interno que corresponte ao id do veículo da frota da empresa.
     * @type {number}
     * @memberof Vehicle
     */
    fleetNumber?: number;
    /**
     * Capacidade do tanque de combustivel do veículo.
     * @type {number}
     * @memberof Vehicle
     */
    fuelTankCapacity?: number;
    /**
     * Quilometragem do veículo na entrada da empresa (no cadastro).
     * @type {number}
     * @memberof Vehicle
     */
    entryMileage?: number;
    /**
     * Média de Quilômetros por litro de conbustivel.
     * @type {number}
     * @memberof Vehicle
     */
    averageConsumption?: number;
    /**
     * Fabricante da carroceria.
     * @type {string}
     * @memberof Vehicle
     */
    bodyManufacturer?: string;
    /**
     * Nome do modelo da carroceria.
     * @type {string}
     * @memberof Vehicle
     */
    bodyModel?: string;
    /**
     * Número de eixos que o veículo possui.
     * @type {number}
     * @memberof Vehicle
     */
    axesNumber?: number;
    /**
     * Descrição do motor.
     * @type {string}
     * @memberof Vehicle
     */
    engineDescription?: string;
    /**
     * Se o veículo possui banheiro
     * @type {boolean}
     * @memberof Vehicle
     */
    hasBathroom?: boolean;
    /**
     * Status operacional atual/desejado do veículo.
     * @type {string}
     * @memberof Vehicle
     */
    status: VehicleStatusEnum;
}


/**
 * @export
 */
export const VehicleStatusEnum = {
    Available: 'available',
    Maintenance: 'maintenance',
    Retired: 'retired'
} as const;
export type VehicleStatusEnum = typeof VehicleStatusEnum[keyof typeof VehicleStatusEnum];